patch slimjab's code
