+++ src/common/__init__.py
@@ -1,1 +1,1 @@
-from src.common.met_office_util import cut_frame, interp_30min
+from src.common.met_office_utils import cut_frame, interp_30min
