--- src/demand/__init__.py
@@ -1,1 +1,0 @@
-from src.demand.demand import get_energy_demand
