+++ src/onsite/__init__.py
@@ -1,0 +1,1 @@
+from src.onsite.onsite import get_energy_demand
